% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{conf}
\alias{conf}
\title{Confidence Probability for Future Effect in Meta-Analysis}
\usage{
conf(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{metaprediction}.}

\item{lower}{Lower bound of the interval (default is \code{-Inf}).}

\item{upper}{Upper bound of the interval (default is \code{Inf}).}
}
\value{
Invisibly returns the confidence probability (a numeric value between 0 and 1).
Prints the result to the console.
}
\description{
Computes the confidence probability that a future outcome \code{theta_new}
lies within a specified interval, based on the predictive distribution from a
random-effects meta-analysis.
}
\seealso{
\code{\link{PredDist}}
}
