% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaeffect.R
\name{remaeffect}
\alias{remaeffect}
\title{Estimate the Average Effect in Random-Effects Meta-Analysis via Edgington's Confidence Distribution}
\usage{
remaeffect(
  es,
  se,
  method = "MC",
  w = rep(1, length(es)),
  level.ci = 0.95,
  B = 100000L,
  seed = NULL,
  mu0 = 0
)
}
\arguments{
\item{es}{Numeric vector of effect estimates from individual studies (length >= 2).}

\item{se}{Numeric vector of standard errors corresponding to each effect estimate (length >= 2).}

\item{method}{Either "MC" for Monte Carlo sampling algorithm or "GAQ" for global adaptive quadrature integration (compare details).}

\item{w}{Study-specific weights.}

\item{level.ci}{Confidence level for the interval estimate (default is 0.95).}

\item{B}{Number of Monte Carlo samples used to estimate the confidence distributions (default is 100,000). Relevant when selecting \code{method = "MC"}.}

\item{seed}{Optional integer to ensure reproducibility of the random sampling. Relevant when selecting \code{method = "MC"}.}

\item{mu0}{Parameter value under the null-hypothesis against which two-sided p-value is computed.}
}
\value{
A list, including not all, but some of these, depending on the method:
\describe{
\item{estimate}{Point estimate of the average effect \eqn{\mu}.}
\item{CI}{Confidence interval for the average effect \eqn{\mu}.}
\item{pval}{Two-sided p-value against H0: \eqn{\mu} = \eqn{\mu_0}.}
\item{cd_mu}{Vector of samples from the confidence distribution of the average effect \eqn{\mu}. Obtained for method \code{MC}}.
\item{cd_tau2}{Vector of samples from the confidence distribution of the between-study heterogeneity \eqn{\tau^2}. Obtained for method \code{MC}}.
\item{fcd}{Marginalized confidence density function of \eqn{\mu}. Obtained for method \code{GAQ}.}
}
}
\description{
Confidence distributions for the average effect \eqn{\mu}, or equivalently, \eqn{p}-value functions, from individual studies
are combined using Edgington's method. The resulting combined confidence distribution of \eqn{\mu},
which is conditional on the heterogeneity parameter \eqn{\tau^2}, is integrated over an approximate confidence distribution of the latter
to account for heterogeneity estimation uncertainty.
}
\details{
This function performs a random-effects meta-analysis using Edgington's confidence distribution: study-specific confidence distributions, respectively,
one-sided p-value functions for the alternative "greater" are constructed from normal pivots unter the normal random-effects model. These
are then combined using Edgington's method, yielding a combined confidence distribution, respectively
combined p-value function, of the average effect \eqn{\mu}. Then, the two approaches proceed as:
\describe{
\item{\code{MC}:}{Generate samples from the approximate confidence distribution of the heterogeneity parameter \eqn{\tau^2},
implied by the generalized heterogeneity statistic (Viechtbauer, 2006). For each sampled \eqn{\tau^{2*}},
generate one \eqn{\mu^*} from Edgington's confidence distribution. Repeatedly sampling heterogeneity  \eqn{\tau^{2*}} and
conditional sampling \eqn{\mu^*}, a confidence distribution of \eqn{\mu}, incorporating uncertainty about the
heterogeneity parameter, is constructed, from which point estimates and confidence intervals are computed.}
\item{\code{GAQ}:}{The confidence density of \eqn{\tau^2}, implied by the generalized heterogeneity statistic,
is constructed by change of variables. Global adaptive quadrature integration is used to integrate Edgington's confidence density
over the confidence density of \eqn{\tau^2}. This typically corresponds to the MC confidence distribution, except if very few
studies are available. In this case, GAQ integration may produce slightly too wide confidence distributions and intervals.}
}

Additionally, study-specific weights can be incorporated, for example to weight studies by
the inverse of their standard errors or variances, or to downweight studies considered to
be at higher risk of bias. When weights other than one are specified, Edgington's weighted
\eqn{p}-value function is used to construct the confidence distribution for the average effect.
}
\examples{
es <- c(0.17,  1.20,  1.10, -0.0019, -2.33)
se <- c(0.52, 0.93, 0.63, 0.3, 0.28)
remaeffect(es = es, se = se, method = "MC")



}
\references{
Viechtbauer, W. (2007). \emph{Confidence intervals for the amount of heterogeneity in meta‐analysi}s. Statistics in medicine, 26(1), 37-52. https://doi.org/10.1002/sim.2514

Held, L., Hofmann, F., & Pawel, S. (2025). \emph{A comparison of combined p-value functions for meta-analysis}. doi:10.1017/rsm.2025.26
}
\author{
David Kronthaler
}
