% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.metaprediction}
\alias{plot.metaprediction}
\title{Plotting Predictive and Confidence Distributions from metaprediction Class Objects}
\usage{
\method{plot}{metaprediction}(obj, param = c("theta_new", "mu", "tau2"), ...)
}
\arguments{
\item{obj}{An object of class \code{metaprediction}.}

\item{param}{The parameter to display. One of \code{theta_new} (predictive distribution), \code{mu} (confidence distribution), or \code{tau2} (confidence distribution).}

\item{...}{Additional arguments passed to \code{graphics::hist()}, including optional \code{main}, \code{xlab}, and \code{ylab} to override default plot title and axis labels.}
}
\description{
Displays the predictive distribution for \code{param = "theta_new"} (future effect from random effects).
For \code{param = "mu"} and \code{param = "tau2"}, the corresponding confidence distributions are shown.
The latter is only valid if the estimation method in \code{PredDist} was not set to \code{"FixedTau2"}.
}
\examples{
es <- c(0.17, 1.20, 1.10, -0.0019, -2.33)
se <- c(0.52, 0.93, 0.63, 0.3, 0.28)
pd <- PredDist(es = es, se = se, method = "FullCD")
plot(pd, param = "theta_new", breaks = 100)
plot(pd, param = "mu", main = "CD for Mu", xlab = "Mu values", ylab = "Density")
}
\author{
David Kronthaler
}
