% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD.R
\name{PredDist}
\alias{PredDist}
\title{Frequentist Predictive Distributions and Prediction Intervals for Meta-Analysis}
\usage{
PredDist(
  es,
  se,
  method = c("FullCD", "SimplifiedCD", "FixedTau2"),
  level.pi = 0.95,
  n_samples = 100000L,
  method.tau2 = "REML"
)
}
\arguments{
\item{es}{Estimates from individual studies (numeric vector, length >= 2).}

\item{se}{Standard errors from individual studies (numeric vector, length >= 2).}

\item{method}{One of "FullCD", "SimplifiedCD", "FixedTau2". Check details for information.}

\item{level.pi}{Level of the prediction interval computed (numeric, between 0 and 1).}

\item{n_samples}{In case method is "FullCD" or "SimplifiedCD", this determines the number of samples generated for the computation of the predictive distribution. For method "FixedTau", this determines the number of samples generated from the deterministic predictive distribution function.}

\item{method.tau2}{In case method is "FixedTau2" or "SimplifiedCD", this determines the method of estimating tau2. Check 'help(meta)' for information.}

\item{...}{Additional arguments handed to 'stats::integrate' (relevant for method "FixedTau2).}
}
\value{
Return a prediction interval.
}
\description{
Frequentist Predictive Distributions and Prediction Intervals for Meta-Analysis
}
\examples{
es <- c(0.17,  1.20,  1.10, -0.0019, -2.33)
se <- c(0.52, 0.93, 0.63, 0.3, 0.28)
PredDist(es = es, se = se, method = "FullCD")
}
