% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PD.R
\name{PredDist}
\alias{PredDist}
\title{Frequentist Predictive Distributions and Prediction Intervals for Meta-Analysis}
\usage{
PredDist(
  es,
  se,
  method = c("FullCD", "SimplifiedCD", "FixedTau2"),
  level.pi = 0.95,
  n_samples = 100000L,
  theta_new = NULL,
  method.tau2 = "REML",
  subdivisions = 300L,
  ...
)
}
\arguments{
\item{es}{Estimates from individual studies (numeric vector, length >= 2).}

\item{se}{Standard errors from individual studies (numeric vector, length >= 2).}

\item{method}{One of "FullCD", "SimplifiedCD", "FixedTau2". Check details for information.}

\item{level.pi}{Level of the prediction interval computed (numeric, between 0 and 1).}

\item{n_samples}{In case method is "FullCD", this determines the number of samples generated for the computation of the predictive distribution.}

\item{theta_new}{In case method is "FixedTau2" or "SimplifiedCD", these are the points for which the predictive density is evaluated and returned.}

\item{method.tau2}{In case method is "FixedTau2", this determines the method of estimating tau2. Check 'help(meta)' for information.}

\item{subdivisions}{Number of subdivisions used for integration.}

\item{...}{Additional arguments handed to 'stats::integrate'.}
}
\value{
Return a prediction interval. In case method is "FullCD" returns a matrix containing samples of theta.new, mu and tau2. In case method is "SimplifiedCD" or "FixedTau2" returns additionally the predictive density in form of a function.
}
\description{
Frequentist Predictive Distributions and Prediction Intervals for Meta-Analysis
}
\examples{
es <- c(0.17,  1.20,  1.10, -0.0019, -2.33)
se <- c(0.52, 0.93, 0.63, 0.3, 0.28)
PredDist(es = es, se = se, method = "FullCD")
}
